+++ packages/web-components/fast-components/src/custom-elements.ts
@@ -10,6 +10,7 @@ import { fastBadge } from "./badge/index";
 import { fastBreadcrumbItem } from "./breadcrumb-item/index";
 import { fastBreadcrumb } from "./breadcrumb/index";
 import { fastButton } from "./button/index";
+import { fastCalendar } from "./calendar/index";
 import { fastCard } from "./card/index";
 import { fastCheckbox } from "./checkbox/index";
 import { fastCombobox } from "./combobox/index";
@@ -80,6 +81,7 @@ export {
     fastBreadcrumb,
     fastBreadcrumbItem,
     fastButton,
+    fastCalendar,
     fastCard,
     fastCheckbox,
     fastCombobox,
@@ -140,6 +142,7 @@ export const allComponents = {
     fastBreadcrumb,
     fastBreadcrumbItem,
     fastButton,
+    fastCalendar,
     fastCard,
     fastCheckbox,
     fastCombobox,