+++ packages/web-components/fast-foundation/src/calendar/index.ts
@@ -1,0 +1,3 @@
+export * from "./calendar";
+export * from "./calendar.template";
+export * from "./date-formatter";