+++ packages/web-components/fast-foundation/src/design-token/custom-property-manager.ts
@@ -6,7 +6,6 @@ import {
     FASTElement,
     observable,
     Observable,
-    TargetedHTMLDirective,
 } from "@microsoft/fast-element";
 
 export const defaultElement = document.createElement("div");
@@ -73,8 +72,10 @@ class HeadStyleElementStyleSheetTarget extends QueuedStyleSheetTarget {
         // there shouldn't exist a case where `sheet` is null,
         // but if-check it just in case.
         if (sheet) {
-            const index = sheet.insertRule(":root{}");
-            this.target = (sheet.rules[index] as CSSStyleRule).style;
+            // https://github.com/jsdom/jsdom uses https://github.com/NV/CSSOM for it's CSSOM implementation,
+            // which implements the DOM Level 2 spec for CSSStyleSheet where insertRule() requires an index argument.
+            const index = sheet.insertRule(":root{}", sheet.cssRules.length);
+            this.target = (sheet.cssRules[index] as CSSStyleRule).style;
         }
     }
 }
@@ -137,8 +138,10 @@ class StyleElementStyleSheetTarget implements PropertyTarget {
         if (sheet) {
             // Safari will throw if we try to use the return result of insertRule()
             // to index the rule inline, so store as a const prior to indexing.
-            const index = sheet.insertRule(":host{}");
-            this.target = (sheet.rules[index] as CSSStyleRule).style;
+            // https://github.com/jsdom/jsdom uses https://github.com/NV/CSSOM for it's CSSOM implementation,
+            // which implements the DOM Level 2 spec for CSSStyleSheet where insertRule() requires an index argument.
+            const index = sheet.insertRule(":root{}", sheet.cssRules.length);
+            this.target = (sheet.cssRules[index] as CSSStyleRule).style;
         } else {
             this.target = null;
         }
@@ -157,13 +160,81 @@ class ElementStyleSheetTarget implements PropertyTarget {
     setProperty(name: string, value: any) {
         DOM.queueUpdate(() => this.target.setProperty(name, value));
     }
+
     removeProperty(name: string) {
         DOM.queueUpdate(() => this.target.removeProperty(name));
     }
 }
 
+/**
+ * Controls emission for default values. This control is capable
+ * of emitting to multiple {@link PropertyTarget | PropertyTargets},
+ * and only emits if it has at least one root.
+ *
+ * @internal
+ */
+export class RootStyleSheetTarget implements PropertyTarget {
+    private static roots = new Set<HTMLElement | Document>();
+    private static properties: Record<string, string> = {};
+    public setProperty(name: string, value: any): void {
+        RootStyleSheetTarget.properties[name] = value;
+
+        for (const target of RootStyleSheetTarget.roots.values()) {
+            PropertyTargetManager.getOrCreate(
+                RootStyleSheetTarget.normalizeRoot(target)
+            ).setProperty(name, value);
+        }
+    }
+
+    public removeProperty(name: string): void {
+        delete RootStyleSheetTarget.properties[name];
+        for (const target of RootStyleSheetTarget.roots.values()) {
+            PropertyTargetManager.getOrCreate(
+                RootStyleSheetTarget.normalizeRoot(target)
+            ).removeProperty(name);
+        }
+    }
+
+    public static registerRoot(root: HTMLElement | Document) {
+        const { roots } = RootStyleSheetTarget;
+        if (!roots.has(root)) {
+            roots.add(root);
+            const target = PropertyTargetManager.getOrCreate(this.normalizeRoot(root));
+            for (const key in RootStyleSheetTarget.properties) {
+                target.setProperty(key, RootStyleSheetTarget.properties[key]);
+            }
+        }
+    }
+
+    public static unregisterRoot(root: HTMLElement | Document) {
+        const { roots } = RootStyleSheetTarget;
+        if (roots.has(root)) {
+            roots.delete(root);
+
+            const target = PropertyTargetManager.getOrCreate(
+                RootStyleSheetTarget.normalizeRoot(root)
+            );
+            for (const key in RootStyleSheetTarget.properties) {
+                target.removeProperty(key);
+            }
+        }
+    }
+
+    /**
+     * Returns the document when provided the default element,
+     * otherwise is a no-op
+     * @param root - the root to normalize
+     */
+    private static normalizeRoot(root: HTMLElement | Document) {
+        return root === defaultElement ? document : root;
+    }
+}
+
 // Caches PropertyTarget instances
-const propertyTargetCache: WeakMap<HTMLElement, PropertyTarget> = new WeakMap();
+const propertyTargetCache: WeakMap<
+    HTMLElement | Document,
+    PropertyTarget
+> = new WeakMap();
 // Use Constructable StyleSheets for FAST elements when supported, otherwise use
 // HTMLStyleElement instances
 const propertyTargetCtor: Constructable<PropertyTarget> = DOM.supportsAdoptedStyleSheets
@@ -176,7 +247,7 @@ const propertyTargetCtor: Constructable<PropertyTarget> = DOM.supportsAdoptedSty
  * @internal
  */
 export const PropertyTargetManager = Object.freeze({
-    getOrCreate(source: HTMLElement): PropertyTarget {
+    getOrCreate(source: HTMLElement | Document): PropertyTarget {
         if (propertyTargetCache.has(source)) {
             return propertyTargetCache.get(source)!;
         }
@@ -184,13 +255,15 @@ export const PropertyTargetManager = Object.freeze({
         let target: PropertyTarget;
 
         if (source === defaultElement) {
+            target = new RootStyleSheetTarget();
+        } else if (source instanceof Document) {
             target = DOM.supportsAdoptedStyleSheets
                 ? new DocumentStyleSheetTarget()
                 : new HeadStyleElementStyleSheetTarget();
-        } else if (isFastElement(source)) {
+        } else if (isFastElement(source as HTMLElement)) {
             target = new propertyTargetCtor(source);
         } else {
-            target = new ElementStyleSheetTarget(source);
+            target = new ElementStyleSheetTarget(source as HTMLElement);
         }
 
         propertyTargetCache.set(source, target);