+++ packages/web-components/fast-foundation/CHANGELOG.md
@@ -1,8 +1,50 @@
 # Change Log - @microsoft/fast-foundation
 
-This log was last generated on Wed, 13 Oct 2021 01:53:37 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Oct 2021 07:11:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.22.1
+
+Thu, 21 Oct 2021 07:11:42 GMT
+
+### Patches
+
+- fix broken tooltip (scomea@microsoft.com)
+
+## 2.22.0
+
+Wed, 20 Oct 2021 07:13:46 GMT
+
+### Minor changes
+
+- add centered positioning option to anchored region (scomea@microsoft.com)
+
+## 2.21.0
+
+Fri, 15 Oct 2021 23:24:14 GMT
+
+### Minor changes
+
+- Implement DesignToken root-element registration for CSS custom property emission of default token values (nicholasrice@users.noreply.github.com)
+
+### Patches
+
+- implement level-2 DOM support for DesignToken in browsers that don't support Constructable StyleSheets (nicholasrice@users.noreply.github.com)
+
+## 2.20.0
+
+Wed, 13 Oct 2021 22:45:16 GMT
+
+### Minor changes
+
+- Adding fast-calendar component (robarb@microsoft.com)
+
+### Patches
+
+- move style assignment for combobox and select into the class and off the DOM node (chhol@microsoft.com)
+- Code fixes for calendar (robarb@microsoft.com)
+- Bump @microsoft/fast-element to v1.6.1 (markwhitfeld@users.noreply.github.com)
 
 ## 2.19.0