+++ sites/site-utilities/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @microsoft/site-utilities
 
-This log was last generated on Wed, 13 Oct 2021 01:53:37 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Oct 2021 07:11:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.0
+
+Thu, 21 Oct 2021 07:11:42 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.14.1 (scomea@microsoft.com)
+
+## 0.9.0
+
+Wed, 20 Oct 2021 07:13:46 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.14.0 (scomea@microsoft.com)
+
+## 0.9.0
+
+Fri, 15 Oct 2021 23:24:14 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.13.1 (nicholasrice@users.noreply.github.com)
+
+## 0.9.0
+
+Wed, 13 Oct 2021 22:45:16 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.13.0 (robarb@microsoft.com)
 
 ## 0.9.0